script/sign: Miniscript support in Tapscript
We make the Satisfier a base in which to store the common methods
between the Tapscript and P2WSH satisfier, and from which they both
inherit.

The PSBT logic is to be able to satisfy pkh() under Tapscript context
(to get the pubkey hash preimage) without wallet data. For instance in
`finalizepsbt` RPC. See also the next commits for a functional test that
exercises this.