miniscript: check maximum stack size during execution
Under Tapscript, due to the lifting of some standardness and consensus
limits, scripts can now run into the maximum stack size during
execution. Any Miniscript that may hit the limit on any of its spending
paths must be marked as unsafe.

To do so we track the maximum stack size at all time during the
execution of a fragment. This is accomplished by:
1. Tracking how the stack is affected *after* execution of a fragment (how
   many items it pushed or popped).
2. Tracking the maximum number of items that may be pushed on the stack
   *during* the execution of a fragment.

We then check the maximum number of items pushed during the execution of
the root fragment with the maximum initial witness size for satisfying
this fragment against the stack size limit. Note this is a conservative
check, since the path that may have the maximum number of items added
during execution may be different than the path that requires the
largest witness. But it should be a pretty close approximation in most
(and non-pathological) cases.

For instance, let's consider an `and_b(X, Y)` fragment (corresponding to
`[X] [Y] BOOLAND`). Let's say the `X` fragment may make the stack bigger
by one item during its execution, but after its executed the stack size
is at most -2 larger.  Let's say the `Y` fragment may grow the stack by
4 elements during its execution and leave the stack with two items less.

In this example the `and_b(X, Y)` fragment will grow the stack size by
at most 2 during its execution (executing `X` then executing `Y`) and
the largest size of the stack after its execution is the initial size
minus 5 items (2 for `X`, 2 for `Y` and 1 for `BOOLAND`).
In absolutes, if the maximum satisfaction size for this fragment is 6,
then the stack size will never grow beyond 8 during its execution.